+++ lib/bsqlite/examples/bsqlite_struct_from_row.rs
@@ -7,6 +7,7 @@
 //! A example that inserts and reads rows with structs that derive [FromRow].
 
 use bsqlite::{Connection, FromRow};
+use const_format::formatcp;
 
 #[derive(FromRow)]
 struct NewPerson {
@@ -46,7 +47,7 @@ fn main() {
     ];
     for person in persons {
         db.execute(
-            format!(
+            formatcp!(
                 "INSERT INTO persons ({}) VALUES ({})",
                 NewPerson::columns(),
                 NewPerson::values()
@@ -56,10 +57,8 @@ fn main() {
     }
 
     // Read rows back
-    let persons = db
-        .query::<Person>(format!("SELECT {} FROM persons", Person::columns()), ())
-        .collect::<Vec<_>>();
-    for person in &persons {
+    let persons = db.query::<Person>(formatcp!("SELECT {} FROM persons", Person::columns()), ());
+    for person in persons {
         println!("{:?}", person);
     }
 }