+++ lib/bsqlite/examples/bsqlite_struct_from_row_chrono.rs
@@ -10,6 +10,7 @@ use std::fmt::Debug;
 
 use bsqlite::{Connection, FromRow};
 use chrono::{DateTime, Utc};
+use const_format::formatcp;
 
 #[derive(FromRow)]
 struct NewPerson {
@@ -18,23 +19,12 @@ struct NewPerson {
     created_at: DateTime<Utc>,
 }
 
-#[derive(FromRow)]
+#[derive(Debug, FromRow)]
 struct Person {
     id: i64,
     name: String,
     age: i64,
     created_at: DateTime<Utc>,
-}
-
-impl Debug for Person {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Person")
-            .field("id", &self.id)
-            .field("name", &self.name)
-            .field("age", &self.age)
-            .field("created_at", &self.created_at.to_string())
-            .finish()
-    }
 }
 
 fn main() {
@@ -65,7 +55,7 @@ fn main() {
     ];
     for person in persons {
         db.execute(
-            format!(
+            formatcp!(
                 "INSERT INTO persons ({}) VALUES ({})",
                 NewPerson::columns(),
                 NewPerson::values()
@@ -75,10 +65,8 @@ fn main() {
     }
 
     // Read rows back
-    let persons = db
-        .query::<Person>(format!("SELECT {} FROM persons", Person::columns()), ())
-        .collect::<Vec<_>>();
-    for person in &persons {
+    let persons = db.query::<Person>(formatcp!("SELECT {} FROM persons", Person::columns()), ());
+    for person in persons {
         println!("{:?}", person);
     }
 }