+++ lib/bsqlite_derive/CHANGELOG.md
@@ -6,9 +6,16 @@ All notable changes to this project will be documented in this file.
 
 _Placeholder_
 
+## [0.1.1] - 2025-02-13
+
+### Changed
+
+-   Make `FromRow` the implementations of the `columns()` and `values()` functions const.
+
 ## [0.1.0] - 2025-02-10
 
 _Initial release_
 
-[Unreleased]: https://github.com/bplaat/crates/compare/bsqlite_derive%2Fv0.1.0...HEAD
+[Unreleased]: https://github.com/bplaat/crates/compare/bsqlite_derive%2Fv0.1.1...HEAD
+[0.1.1]: https://github.com/bplaat/crates/releases/tag/bsqlite_derive%2Fv0.1.1
 [0.1.0]: https://github.com/bplaat/crates/releases/tag/bsqlite_derive%2Fv0.1.0